Fix hand evaluation, chip accounting, bot logic and simulation math
Correctness fixes across the evaluator, game loop, bot profiles and the
Monte Carlo / betting mathematics. All findings were reproduced before
being fixed and are covered by regression tests (56 passing).

Hand evaluation:
- isStraight returned the first run of five rather than the highest, so
  six-card runs scored a rank low and a royal flush was reported as a
  king-high straight flush.
- Four-of-a-kind and two-pair kickers ignored paired cards, so AAAA-KK-Q
  played the queen and AA-KK-QQ-J played the jack.
- Removed model/hand_evaluator.cpp, dead code in neither SRC nor LIB_SRC.

Chip accounting:
- There was no pot. Showdown paid a hardcoded 200 regardless of what was
  staked, destroying chips on big pots and creating them on early folds.
- Player::bet silently no-opped when short-stacked; it now returns the
  amount actually wagered, clamped to the stack as an all-in.
- Uncalled portions of a bet are returned to the bettor.

Bot behaviour:
- The bot could only call or fold, so it checked back made hands.
- HardPlus priced every decision off a hardcoded 200 pot / 100 call, and
  its inline simulation never completed the board.
- Fixed an out-of-bounds read in hasStraightDraw reachable at every
  difficulty, and its gutshot test that only matched open-ended draws.

Game loop:
- Deck reseeded from time(nullptr) each round, so rounds starting in the
  same second dealt identical cards.
- Difficulty input fell back to Medium silently on any typo, uppercase, or
  buffered input left by typing "skip" at the viewer prompt.
- Spinner indexed a 3-element array with `i & 4`, reading out of bounds.
- GameLogger was fully implemented but never called.

Simulation and betting math:
- getConfidenceInterval ignored levels below 90%, returning a 95% interval;
  now derives z from the requested level via the probit function.
- Confidence intervals were computed on win rate while decisions are made
  on equity; added an equity interval with the correct three-valued
  variance. Empirical coverage of equity went from 18% to 95%.
- kellyFraction returned 0 at p >= 1.0, advising "don't bet" on the nuts.

Verified against exact enumeration of all 990 opponent holdings and
against the symmetry invariant that mean equity over random hands is 0.5.

+++ animation/card_animation.cpp
@@ -65,8 +65,10 @@ void CardAnimation::dealCommunityCards(const std::vector<Card> &cards, const std
     std::cout << "\n" << BOLD << stageColor << "╔════════════════════════════════════════════════════╗" << RESET << "\n";
     std::cout << BOLD << stageColor << "║" << RESET << "                  " << BOLD << stageColor << "🎴 " << stage << " 🎴" << RESET;
     
-    // Add spacing to center the text
-    int padding = 52 - 23 - stage.length();
+    // Add spacing to center the text. stage.length() is unsigned, so the
+    // subtraction has to be done in a signed type - otherwise a long stage
+    // name wraps to a huge positive count instead of going negative.
+    int padding = 52 - 23 - static_cast<int>(stage.length());
     for (int i = 0; i < padding; i++) std::cout << " ";
     std::cout << BOLD << stageColor << "║" << RESET << "\n";
     std::cout << BOLD << stageColor << "╚════════════════════════════════════════════════════╝" << RESET << "\n";
